clean up `convert`, `decryption`, and `pod` modules
