+++ zk-token-sdk/src/instruction/ctxt_ctxt_equality.rs
@@ -25,7 +25,7 @@ use {
 use {
     crate::{
         instruction::{ProofType, ZkProofData},
-        pod,
+        zk_token_elgamal::pod,
     },
     bytemuck::{Pod, Zeroable},
 };
