+++ zk-token-sdk/src/instruction/transfer_with_fee.rs
@@ -29,7 +29,7 @@ use {
 use {
     crate::{
         instruction::{ProofType, ZkProofData},
-        pod,
+        zk_token_elgamal::pod,
     },
     bytemuck::{Pod, Zeroable},
 };
