+++ zk-token-sdk/src/instruction/withdraw.rs
@@ -16,7 +16,7 @@ use {
 use {
     crate::{
         instruction::{ProofType, ZkProofData},
-        pod,
+        zk_token_elgamal::pod,
     },
     bytemuck::{Pod, Zeroable},
 };
