+++ zk-token-sdk/src/instruction/zero_balance.rs
@@ -18,7 +18,7 @@ use {
 use {
     crate::{
         instruction::{ProofType, ZkProofData},
-        pod,
+        zk_token_elgamal::pod,
     },
     bytemuck::{Pod, Zeroable},
 };
