--- zk-token-sdk/src/zk_token_elgamal/convert.rs
@@ -1,119 +1,0 @@
-pub use target_arch::*;
-use {super::pod, crate::curve25519::ristretto::PodRistrettoPoint};
-
-#[cfg(not(target_os = "solana"))]
-mod target_arch {
-    use {
-        super::pod,
-        crate::{
-            curve25519::scalar::PodScalar,
-            errors::ProofError,
-            instruction::{
-                transfer::{TransferAmountEncryption, TransferPubkeys},
-                transfer_with_fee::{FeeEncryption, FeeParameters, TransferWithFeePubkeys},
-            },
-        },
-        curve25519_dalek::{ristretto::CompressedRistretto, scalar::Scalar},
-        std::convert::TryFrom,
-    };
-
-    impl From<Scalar> for PodScalar {
-        fn from(scalar: Scalar) -> Self {
-            Self(scalar.to_bytes())
-        }
-    }
-
-    impl TryFrom<PodScalar> for Scalar {
-        type Error = ProofError;
-
-        fn try_from(pod: PodScalar) -> Result<Self, Self::Error> {
-            Scalar::from_canonical_bytes(pod.0).ok_or(ProofError::CiphertextDeserialization)
-        }
-    }
-
-    impl From<CompressedRistretto> for pod::CompressedRistretto {
-        fn from(cr: CompressedRistretto) -> Self {
-            Self(cr.to_bytes())
-        }
-    }
-
-    impl From<pod::CompressedRistretto> for CompressedRistretto {
-        fn from(pod: pod::CompressedRistretto) -> Self {
-            Self(pod.0)
-        }
-    }
-}
-
-#[cfg(target_os = "solana")]
-#[allow(unused_variables)]
-mod target_arch {}
-
-#[cfg(test)]
-mod tests {
-    use {
-        super::*,
-        crate::{encryption::pedersen::Pedersen, range_proof::RangeProof},
-        merlin::Transcript,
-        std::convert::TryInto,
-    };
-
-    #[test]
-    fn test_pod_range_proof_64() {
-        let (comm, open) = Pedersen::new(55_u64);
-
-        let mut transcript_create = Transcript::new(b"Test");
-        let mut transcript_verify = Transcript::new(b"Test");
-
-        let proof = RangeProof::new(vec![55], vec![64], vec![&open], &mut transcript_create);
-
-        let proof_serialized: pod::RangeProof64 = proof.try_into().unwrap();
-        let proof_deserialized: RangeProof = proof_serialized.try_into().unwrap();
-
-        assert!(proof_deserialized
-            .verify(vec![&comm], vec![64], &mut transcript_verify)
-            .is_ok());
-
-        // should fail to serialize to pod::RangeProof128
-        let proof = RangeProof::new(vec![55], vec![64], vec![&open], &mut transcript_create);
-
-        assert!(TryInto::<pod::RangeProof128>::try_into(proof).is_err());
-    }
-
-    #[test]
-    fn test_pod_range_proof_128() {
-        let (comm_1, open_1) = Pedersen::new(55_u64);
-        let (comm_2, open_2) = Pedersen::new(77_u64);
-        let (comm_3, open_3) = Pedersen::new(99_u64);
-
-        let mut transcript_create = Transcript::new(b"Test");
-        let mut transcript_verify = Transcript::new(b"Test");
-
-        let proof = RangeProof::new(
-            vec![55, 77, 99],
-            vec![64, 32, 32],
-            vec![&open_1, &open_2, &open_3],
-            &mut transcript_create,
-        );
-
-        let proof_serialized: pod::RangeProof128 = proof.try_into().unwrap();
-        let proof_deserialized: RangeProof = proof_serialized.try_into().unwrap();
-
-        assert!(proof_deserialized
-            .verify(
-                vec![&comm_1, &comm_2, &comm_3],
-                vec![64, 32, 32],
-                &mut transcript_verify,
-            )
-            .is_ok());
-
-        // should fail to serialize to pod::RangeProof64
-        let proof = RangeProof::new(
-            vec![55, 77, 99],
-            vec![64, 32, 32],
-            vec![&open_1, &open_2, &open_3],
-            &mut transcript_create,
-        );
-
-        assert!(TryInto::<pod::RangeProof64>::try_into(proof).is_err());
-    }
-}
